lab9-incident-response/backend/requirements.txt
flask==2.3.3
prometheus_client==0.17.1
psycopg2-binary==2.9.7
redis==4.6.0